JPA delete beer and customer by id Not Found - test

+++ demo/demo/src/main/java/it/johnson/demo/service/CustomerServiceJPA.java
@@ -63,8 +63,12 @@ public class CustomerServiceJPA implements CustomerService{
     }
 
     @Override
-    public void deleteById(UUID customerId) {
-        customerRepository.deleteById(customerId);
+    public Boolean deleteById(UUID customerId) {
+        if(customerRepository.existsById(customerId)){
+            customerRepository.deleteById(customerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
